Backend/requirements.txt
flask
werkzeug
openai-whisper
torch
transformers
moviepy
better_profanity
flask-cors
ffmpeg-python
